first/requirements.txt
astroid==2.3.3
Django==2.2.12
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.7
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
requests==2.11.1
